utils: move retry helpers to pkg/utils to break import cycle
